make individual files for incantations make player class in game.py
