+++ src/incantation/fireball.py
@@ -1,0 +1,6 @@
+from src.incantation.incantation import Incantation
+
+
+class Fireball(Incantation):
+    def cast(self, target):
+        target.hp -= 10
