+++ src/incantation/heal.py
@@ -1,0 +1,6 @@
+from src.incantation.incantation import Incantation
+
+
+class Heal(Incantation):
+    def cast(self, target):
+        target.hp += 5
