+++ src/incantation/incantation.py
@@ -5,18 +5,3 @@ class Incantation(ABC):
     @abstractmethod
     def cast(self, target):
         pass
-
-
-class Fireball(Incantation):
-    def cast(self, target):
-        target.hp -= 10
-
-
-class Heal(Incantation):
-    def cast(self, target):
-        target.hp += 5
-
-
-class SuperHeal(Incantation):
-    def cast(self, target):
-        target.hp += 50
