+++ src/incantation/superheal.py
@@ -1,0 +1,6 @@
+from src.incantation.incantation import Incantation
+
+
+class SuperHeal(Incantation):
+    def cast(self, target):
+        target.hp += 50
